Remove old node helpers, move remaining code to nodes/utils.py
Delete molecularnodes/nodes/nodes.py and node_management.py. The still-used
pieces (custom_*_iswitch, style mappings, swap, get_star_node,
get_final_style_nodes, remove_style_node) move to molecularnodes/nodes/utils.py;
the tree-walking helpers only the old tests used are removed, and those tests
now use the nodebpy tree.reset() API instead.

Also remove blender.mesh.evaluate_using_mesh: densities and tests read
evaluated geometry through databpy.GeometrySet directly.

get_final_style_nodes now matches 'Style' anywhere in the node name (as the UI
styles list already did) so density style nodes are found too.
