+++ molecularnodes/blender/mesh.py
@@ -2,38 +2,7 @@ import bpy
 import numpy as np
 from databpy.attribute import AttributeTypes
 from databpy.object import create_bob
-from ..nodes import nodes
 from . import coll
-
-
-def evaluate_using_mesh(obj: bpy.types.Object) -> bpy.types.Object:
-    """
-    Evaluate the object using a debug object. Some objects can't currently have their
-    Geometry Node trees evaluated (such as volumes), so we source the geometry they create
-    into a mesh object, which can be evaluated and tested.
-
-    Parameters
-    ----------
-    object : bpy.types.Object
-        The object to be evaluated.
-
-    Returns
-    -------
-    bpy.types.Object
-
-    Notes
-    -----
-    Intended for debugging only.
-    """
-    # create an empty mesh object. It's modifiers can be evaluated but some other
-    # object types can't be currently through the API
-    bob = create_bob()
-    mod = nodes.get_mod(bob.object)
-    mod.node_group = nodes.create_debug_group()
-    mod.node_group.nodes["Object Info"].inputs["Object"].default_value = obj
-
-    # need to use 'evaluate' otherwise the modifiers won't be taken into account
-    return bob.evaluate()
 
 
 def create_data_object(
