+++ molecularnodes/entities/molecule/base.py
@@ -28,7 +28,7 @@ from ...converters import universe_from_atoms
 from ...material import PresetMaterial, append_material
 from ...nodes import handlers as node_handlers
 from ...nodes.geometry import AssemblyInstance
-from ...nodes.nodes import STYLE_LITERALS, STYLE_NODE_MAPPING
+from ...nodes.utils import STYLE_LITERALS, STYLE_NODE_MAPPING
 from ...utils import _UNSET, count_value_changes, temp_override_property
 from ..base import EntityType, MolecularEntity, MolecularTree
 from ..utilities import (
