+++ molecularnodes/nodes/__init__.py
@@ -1,15 +1,13 @@
 from . import (
     geometry,
     handlers,
-    node_management,
-    nodes,
     shader,
+    utils,
 )
 
 __all__ = [
-    "nodes",
-    "handlers",
-    "node_management",
     "geometry",
+    "handlers",
     "shader",
+    "utils",
 ]
